Removed premptive authentication with LDAP before returning adapter.
Previously, the code authenticated using LDAP first to "try out" whether LDAP authentiation works before returning the LDAP adapter in the login filter. This is a problem if your university uses 2-factor authentication, because it doubles the number of 2-factor authentication push requests that must be responded to in order to log in. My opinion is that we should just keep it simple, and LDAP auth should either be "on" or "off".