Store LDAP settings in ini file, reduce LDAP binds. (#14)
LDAP settings may now be stored in the [ldap] section of a central_auth.ini file, so there is an alternative to storing credentials in the database.

The result of the LDAP authentication is now stored between calls, to prevent the Zend LDAP Adapter from authenticating twice. The Zend LDAP adapter still binds twice, but this prevents that from being doubled to 4 calls.